[example-code] LZ77: - Using int as indices larger files can be encoded but compression is decreased. - Using binary I/O. Fixes garbage at end of files.

+++ trunk/example-code/examples/LZ77/LZ77.h
@@ -6,13 +6,13 @@ using std::vector;
 
 struct EncodingPair
 {
-	short int Reference;
+	int Reference;
 	char Character;
 };
 
 struct DictionaryEntry
 {
-	short int Reference;
+	int Reference;
 	vector<char> Phrase;
 	EncodingPair Encoding;
 };
